statistics afgesplitst klaar
